+++ src/adminTool/quadtree/CollisionlessQuadtree.java
@@ -10,7 +10,7 @@ public class CollisionlessQuadtree {
     private final IntList elements;
 
     public CollisionlessQuadtree(final int elements, final int maxHeight, final IQuadtreePolicy qp,
-            final ICollisionPolicy cp, final int size) {
+            final ICollisionPolicy cp, final double size) {
         this.elements = new IntList();
 
         for (int element = 0; element < elements; element++) {
@@ -23,12 +23,12 @@ public class CollisionlessQuadtree {
     }
 
     private void add(final int element, final int maxHeight, final IQuadtreePolicy qp, final ICollisionPolicy cp,
-            final int x, final int y, final int height, final int size) {
+            final double x, final double y, final int height, final double size) {
         final int destHeight = getCollisionlessHeight(element, qp, cp, x, y, height, size);
         add(element, x, y, size, height, destHeight, maxHeight, qp);
     }
 
-    private void add(final int element, final int x, final int y, final int size, final int height,
+    private void add(final int element, final double x, final double y, final double size, final int height,
             final int destHeight, final int maxHeight, final IQuadtreePolicy qp) {
         if (qp.intersects(element, height, x, y, size)) {
             if (height >= destHeight) {
@@ -42,11 +42,11 @@ public class CollisionlessQuadtree {
                         children[i] = new CollisionlessQuadtree();
                     }
                 }
-                final int nSize = size / 2;
+                final double nSize = size / 2;
                 final int nHeight = height + 1;
                 for (int i = 0; i < children.length; i++) {
-                    final int nx = x + (i % 2) * nSize;
-                    final int ny = y + (i / 2) * nSize;
+                    final double nx = x + (i % 2) * nSize;
+                    final double ny = y + (i / 2) * nSize;
                     children[i].add(element, nx, ny, nSize, nHeight, destHeight, maxHeight, qp);
                 }
             }
@@ -54,17 +54,17 @@ public class CollisionlessQuadtree {
     }
 
     private int getCollisionlessHeight(final int element, final IQuadtreePolicy qp, final ICollisionPolicy cp,
-            final int x, final int y, final int height, final int size) {
+            final double x, final double y, final int height, final double size) {
         if (isLeaf() || !qp.intersects(element, height, x, y, size) || !intersects(element, cp, height)) {
             return height;
         }
 
         int ret = height + 1;
-        final int nSize = size / 2;
+        final double nSize = size / 2;
         final int nHeight = height + 1;
         for (int i = 0; i < children.length; i++) {
-            final int nx = x + (i % 2) * nSize;
-            final int ny = y + (i / 2) * nSize;
+            final double nx = x + (i % 2) * nSize;
+            final double ny = y + (i / 2) * nSize;
 
             ret = Math.max(children[i].getCollisionlessHeight(element, qp, cp, nx, ny, nHeight, nSize), ret);
         }
